+ config/PriorRTOSConfig: Moved config from source to config folder. * source/kernel/: Fixed build errors.
